rename: 전체 파일명 변경
폴더와 폴더 안의 이름이 겹쳐 파일 이름을 index.ts로 변경함
root의 index.ts 경로 더 간단하게 수정
@/component -> @comp로 수정
@/views -> @views로 수정
