+++ src/components/Footer/index.ts
@@ -1,4 +1,5 @@
-import styles from "./footer.css";
+import styles from "./index.css";
+
 function Footer() {
   return `
 <footer class=${styles["footer"]}>
